rename file in components
