+++ src/components/Avatar/style.css
