+++ src/components/Icon/style.css
